Addition of notification configuration and pagerduty provider.
The notifier package has been added and setup in a provider
manner so that additional notification backends can be added
over time. Currently PagerDuty is the first and only notifier
backend has been setup, and will likely change when this is
integrated properly and called.

The CLI has been updated to reflect the additional params.