Fix incorrect cluster_scaling and job_scaling enabled config keys.
When a rework of the configuration parsing was done to accomodate
CLI falgs both the cluster_scaling and job_scaling enabled flags
we wrongly prefixed with their top level key. This meant that the
config files params allowed did not match the config struct
mapstructure designation and so was failing to work as expected.

This updates the config_parse to check for the correct keys thus
allowing enabled to be set on cluster and job scaling.