Support Persistent State Tracking
This commit provides enhancements to the state
tracking mechanism including support for
persistent storage in a Consul Key/Value store.

When Replicator begins evaluating if a scaling
operation is required, it will first look for
persistent state tracking information and load
it into memory. If no persistent state tracking
information is found, or the attempt to retrieve
it fails, the in-memory data will be used.

This is critical to allowing Replicator to run as
a Nomad job or docker container while also
ensuring we can pick up where we left off. This
also clears the way for persistently locking the
daemon in `failsafe` mode after certain failure
modes are encountered (see #86 for more details).

Closes #83.