Modify consul key location to have base path 'replicator/config'.
Replicators top level consul_key_value parameter now defaults to
replicator/config and is intended to be used as a storage area
for both persistant state, leader locking and job scaling docs.

The GetJobScalingPolicies function now suffixes /jobs to the config
directive and is the location where scaling documents should be
stored. It could be that we make this parameter configurable later
but I was conscious off adding too many configuration params to
replicator at this point.

This commit also updates the README to point to the Replicator wiki
page so that we do not have duplicate information.

Closes #94