Replicator now runs leadership locking using Consul sessions +  KV
This requirement helps move Replicator towards being able to run
as a Nomad job. Each replicator instance will obtain a Consul
session and then attempt to gain a lock on a Consul KV. The session
is continually renewed in the background and non-leadership locking
instances will continue to attempt to gain the lock.

If a leader instance is shutdown, it will attempt to release its
session and lock allowing other instances to obtain the lock quicker
than waiting for the TTL to expire.

Telemetry has been added as a counter for leadership aquisition
events.

Closes #32