Evaluate cluster-node count against asg desired before scaling.
This change brings in an evaluation between the desired ASG count
and the Nomad worker pool before scaling the cluster out. This
stops us incramenting the ASG count further if we have no need to
as Replicator tracks node launches and failures to join the
cluster.

ScaleOut will now also perform a check against the ASG MaxSize
before incramenting the count; as if the incramentation were to
violate the MaxSize the AWS call would fail anyway.

Closes #70