Protect Nomad Agent Node Running Replicator Leader
This commit adds intelligence to Replicator to
determine if it is running as a Nomad job and if
so, to attempt to identify the agent node on which
it is running.

If the node is identified, it is registered in the
state tracking information as a protected node.
The least-allocated node discovery method has
been modified to exclude the protected node during
the discovery process.

This feature is only used by the elected Replicator
leader. If this feature encounters errors, the
original behavior is retained.

Closes #119.