Introduce Node Discovery With Dynamic Refresh
This commit introduces a new mechanism for
discovering worker pools and their associated
nodes.

The new mechanism introduces a node registry, a
shared object for tracking pools of Nomad agents.
Nodes are discovered by implementing blocking
Nomad queries to continuously watch for changes
in node membership and register or deregister the
nodes.

This commit also migrates the cluster scaling
configuration to meta parameters within the Nomad
agent configuration.

A node is considered valid for registration if the
following conditions are met:
- The node has all required Replicator meta
configuration parameters.
- Scaling is enabled by setting
`replicator_enabled` to `true`.
- The nodes is not in drain mode.

A node is deregistered if the following conditions
are met:
- The node is not in a down state.
- The node is placed in drain mode.
- Scaling is disabled by setting
`replicator_enabled` to `false`.
- Any or all required Replicator meta configuration
parameters are missing.

If a worker pool is left with no registered nodes,
the worker pool is also deregistered.

Closes TIOCE-254
Closes TIOCE-253