Fix an issue with job/node watcher retry upon API call failure.
Upon connectivity issues to Nomad, the job/node watcher uses a
return once the error is logged which causes the watcher to
stop working even after connectivity is restored.

This change modifies the behaviour to sleep and continue so that
when connectivity is lost and restored the watchers can continue
to work as expected.