Fix bug in session renew with intermittent Consul connectivity.
If Replicator were to lose connectivity to Consul which was then
re-established, the session stored in Replicators internal state
would have expired in Consul meaning Replicator could no longer
even attempt for leadership election.

This fix puts in place a check and revoke if the Consul session has
expired which allows intermittent Consul connectivity to not affect
Replicator ablity to attempt to gain leadership.