Fixes bug in job_scaling where purged jobs cause panic on js run.
The bug is triggered because when a Nomad job is purged, the job
policy watcher will not fire as the job dissapears from the Nomad
job info API. There is therefore no way to tell and remove the
scaling policy entry for the job, and on the next Replicator js
run we are unabe to find the job and therefore panic.

The fix puts in better error handling around the
GetTaskGroupResources and EvaluateJobScaling functions as well as
implaments a mthod to tidy up job policies for purged jobs.

Note - github.com/hashicorp/nomad/issues/1849 has heavy baring on
this change, and if ever that issue is resolved, it is likely this
purge functionality will need to change.