Add Initialization Flag To State Read Method
This commit introduces a boolean flag to the `ReadState` method to
allow a caller to specify the desired initialization behavior when
no existing state object is found in Consul.

This allows admin CLI tools that accept a state path to detect
when no existing state is found and throw an error.

Closes #159