Fix Scaling Deadlock Due To Leader Election
This commit fixes an intermittent deadlock issue caused by the leader
setting an explicit false assertion at the beginning of the leader
election method. This results in the leader temporarily representing
itself as not having leadership to the scaling tickers.

This fix removes the explicit false assertion and sets a slight
increase in the lock TTL.

Closes #177