Reset Failure Count When Exiting Failsafe
This commit ensures the failure count of a scalable resource is
reset when an operator resets the failsafe circuit breaker.

This resolves a bug that results in the failsafe circuit breaker
always being tripped for a scalable resource at the next scaling
evaluation, if the circuit breaker was tripped automatically -
that is, not by an operator.

A test has been added to prove the fix. Also includes small
fixes in the `failsafe` command help verbage.

Closes #188.