Improve accuracy and detail of job_scaling telemetry metrics.
This change imrpoves the accuracy and detail of job_scaling
telemetry. Previously a counter would be sent upon every scaling
request regardless of the success of the resulting deployment.

This change introduces better metric logic and takes into account
the scaling deployment outcome to send eitehr a success or failure
namespace divider on the metric name. This allows total requests
to be tracked as well as the number of failures and successes.