Improve Node Protection Mechanism
If Replicator wins the leader election, it automatically determines
if it is running as a Nomad job and attempts to protect the worker
pool node hosting its allocation.

If a node is marked as protected, Replicator will consider it
ineligible when identifying a node for scale in.

This commit improves the node protection mechanism, provides some
organization and ensures the `LeastAllocatedNode` method in the
`client/nomad` package actually uses this value during evaluation.

Closes #197.