Prevent Processing of Nomad System Jobs
This commit prevents Replicator from registering a job as scalable
if the job type is `system`.

Before this fix, if a user specified a scaling policy in a system job,
Replicator would register the job as scalable. This results in a SIGSEV
when Replicator attempts to check the deployment status of the job.

Closes #215.