Use `AllowStale: true` in all JobScaling activies where its safe.
In order to lower the pressure Replicator puts on the Nomad leader
this change introduces adding the QueryOption `AllowStale: true`
to all calls which are safe to do so. The staleness as stated in
the Nomad docs will be less than 50ms, therefore this should not
have any negative impact in the running of Replicator.

Closes #224