Make Async Job Scaling Block On Concurrent Scaling
This commit makes the asynchronous job scaling method block on
completion of all concurrent job scaling threads.

This change is part of a restructuring of the scaling concurrency
model to reduce the number of concurrent scaling operations that
Replicator initiates. In clusters with a large number of jobs, the
time it takes to complete job scaling operations might exceed the
scaling interval. This change ensures we finish all scaling
operations before returning to the runner.