Addition of API framework and required configuration.
This change add a HTTP and RPC framework which binds to a
specified address and user specified ports, with defaults for all
available. Replicator when started will therefore run a server
listening for requests.

This change does not include any endpoints; these will be raised
as seperate requests as they are likely to involve a non-trival
amount of work.

Closes #186