Only bind RPC to localhost interface on static port for now.
RPC is unable to bind to 0.0.0.0 as it is not advertisable; and as
the RPC interface is only currently used by the local Replicator
agent we can leave this static for now until other features come
in. This allows the -bind-address to be 0.0.0.0 so Replicator run
can properly on Nomad.